+++ comet/utility/test/test_event_db.py
@@ -1,10 +1,15 @@
 # Comet VOEvent Broker.
 # Event database tests.
 
-import tempfile
+import os
 import shutil
+import stat
+import tempfile
+import time
+from functools import reduce
+from itertools import repeat, permutations
 from multiprocessing.pool import ThreadPool
-from itertools import repeat
+from operator import __or__
 
 from twisted.trial import unittest
 
@@ -16,6 +21,35 @@ class Event_DB_TestCase(unittest.TestCase):
         self.event_db_dir = tempfile.mkdtemp()
         self.event_db = Event_DB(self.event_db_dir)
         self.event = DummyEvent()
+
+    def test_non_existing_dir(self):
+        # If the root for the Event_DB doesn't exist, we should create it.
+        # Note the relative path means that this DB will be created in the
+        # _trial_temp directory.
+        event_db = Event_DB("event_db_test_%.5f" % (time.time(),))
+        self.assertTrue(event_db.check_event(self.event))
+
+    def test_dir_is_file(self):
+        # If the path specified for the Event_DB *does* exist but isn't a
+        # directory, then we should fail fast.
+        filename = "event_db_test_%.5f" % (time.time(),)
+        open(filename, 'w').close()
+        self.assertRaises(RuntimeError, Event_DB, filename)
+
+    def test_dir_is_unusable(self):
+        # If the path specified for the Event_DB exists and is a directory,
+        # but we don't have permissions to use it, fail fast.
+        filename = "event_db_test_%.5f" % (time.time(),)
+        os.makedirs(filename)
+        os.chmod(filename, 0)
+        self.assertRaises(RuntimeError, Event_DB, filename)
+        for n_perms in [1, 2]:
+            for perms in permutations([stat.S_IRUSR, stat.S_IWUSR, stat.S_IXUSR],
+                                      n_perms):
+                os.chmod(filename, reduce(__or__, perms))
+                self.assertRaises(RuntimeError, Event_DB, filename)
+        os.chmod(filename, stat.S_IRUSR | stat.S_IWUSR | stat.S_IXUSR)
+        self.assertTrue(Event_DB(filename).check_event(self.event))
 
     def test_unseen(self):
         # Unseen event -> return True